+++ main.go
@@ -47,7 +47,7 @@ func (rr *replReader) ReadLine() (string, bool) {
 // Version is bumped at release time. Override at build with:
 //
 //	go build -ldflags "-X main.Version=v0.2.0"
-var Version = "v0.32.0"
+var Version = "v0.33.0"
 
 const (
 	cReset  = "\033[0m"